TASK - 3456(Added default service icon and service page budget calculations)

- Added default service icon on AddServiceAdapter.kt file.
- Implemented budget value not greater than 10 digit validation on event and service details pages.

Related work items: #3456

+++ app/src/main/java/com/smf/customer/data/model/request/EventInfoDTO.kt
@@ -1,5 +1,7 @@
 package com.smf.customer.data.model.request
 
+import java.math.BigDecimal
+
 data class EventInfoDTO(
     val eventId: Int,
     val eventMetaDataDto: EventMetaDataDto,
@@ -13,7 +15,7 @@ data class EventInformationDto(
     val actualEventBudget: String,
     val attendeesCount: String,
     val currencyType: String,
-    val estimatedEventBudget: Int,
+    val estimatedEventBudget: BigDecimal,
     val eventDate: String,
     val eventName: String
 )
